+++ packages/app/src/index.ts
@@ -1,3 +1,4 @@
 export type { TypesLocales, TypesTimeZones } from '@zerodep/types';
 
+export * from '@zerodep/locale';
 export * from '@zerodep/utils';