+++ packages/format/format/src/index.ts
@@ -1,0 +1,4 @@
+export type { ZeroDepError } from '@zerodep/errors';
+
+export * from '@zerodep/format.currency';
+export * from '@zerodep/format.errors';