+++ packages/locale/locale.get/src/lib/getIntlLocales.spec.ts
@@ -1,0 +1,7 @@
+import { getIntlLocales } from './getIntlLocales';
+
+describe('getIntlLocales', () => {
+  it('should return an array of one or more locales', () => {
+    expect(getIntlLocales().length).toBeGreaterThanOrEqual(1);
+  });
+});