+++ packages/locale/locale.get/src/lib/getNodeLocales.spec.ts
@@ -1,0 +1,7 @@
+import { getNodeLocales } from './getNodeLocales';
+
+describe('getNodeLocales', () => {
+  it('should return an array of one or more locales', () => {
+    expect(getNodeLocales().length).toBeGreaterThanOrEqual(1);
+  });
+});