+++ packages/types/types/src/index.ts
@@ -1,2 +1,3 @@
+export type { TypesCurrencies } from '@zerodep/types.currencies';
 export type { TypesLocales } from '@zerodep/types.locales';
 export type { TypesTimeZones } from '@zerodep/types.timezones';