+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
 export * from '@zerodep/errors';
 export * from '@zerodep/can';
+export * from '@zerodep/format';
 export * from '@zerodep/guard';
 export * from '@zerodep/is';
 export * from '@zerodep/to';